+++ R/table_getNearestLocation.R
@@ -11,8 +11,10 @@
 #' 
 #' @return Tibble of known locations.
 #' 
-#' @description Returns a tibble of known locations from \code{locationTbl}, one 
-#' for each incoming location. If no known location is found for a particular
+#' @description Returns a tibble of the known locations from \code{locationTbl} 
+#' that are closest to the vector of target locations specified by \code{longitude}
+#' and \code{latitude}. Only a single known location is returned for each
+#' incoming target location. If no known location is found for a particular
 #' incoming location, that record in the tibble will contain all \code{NA}.
 #' 
 #' @examples