+++ R/table_leaflet.R
@@ -90,8 +90,8 @@ table_leaflet <- function(
   
   hasCoreMetadata <- all(coreMetadataNames %in% names(locationTbl))
   
-  if ( is.null(jitter) || is.na(jitter) ) {
-    jitter <- 0
+  if ( is.null(jitter) || is.na(jitter) || jitter == 0 ) {
+    jitter <- 1e-9
   }
 
   # * argsList -----